Fixed bug that assumed size_t == unsigned long.

+++ src/vector_integer/tmpl_create_empty_short_vector.c
@@ -75,8 +75,7 @@
 #include <libtmpl/include/tmpl_vector_integer.h>
 
 /*  Function for allocating memory to a short vector.                         */
-tmpl_ShortVector *
-tmpl_Create_Empty_ShortVector(unsigned long int length)
+tmpl_ShortVector *tmpl_Create_Empty_ShortVector(size_t length)
 {
     /*  Declare a variable for the vector pointer we're returning.            */
     tmpl_ShortVector *vec;
@@ -95,7 +94,7 @@ tmpl_Create_Empty_ShortVector(unsigned long int length)
     vec->length = length;
 
     /*  If the length is zero, set the data to a NULL pointer.                */
-    if (length == 0UL)
+    if (length == (size_t)0)
     {
         vec->data = NULL;
         return vec;
@@ -108,7 +107,7 @@ tmpl_Create_Empty_ShortVector(unsigned long int length)
     if (vec->data == NULL)
     {
         vec->error_occurred = tmpl_True;
-        vec->length = 0UL;
+        vec->length = (size_t)0;
         vec->error_message = tmpl_strdup(
             "Error Encountered: libtmpl\n"
             "\tFunction Name: tmpl_Create_Empty_ShortVector\n\n"
@@ -123,8 +122,7 @@ tmpl_Create_Empty_ShortVector(unsigned long int length)
 /*  End of tmpl_Create_Empty_ShortVector.                                     */
 
 /*  Function for allocating memory to an unsigned short vector.               */
-tmpl_UShortVector *
-tmpl_Create_Empty_UShortVector(unsigned long int length)
+tmpl_UShortVector *tmpl_Create_Empty_UShortVector(size_t length)
 {
     /*  Declare a variable for the vector pointer we're returning.            */
     tmpl_UShortVector *vec;
@@ -143,7 +141,7 @@ tmpl_Create_Empty_UShortVector(unsigned long int length)
     vec->length = length;
 
     /*  If the length is zero, set the data to a NULL pointer.                */
-    if (length == 0UL)
+    if (length == (size_t)0)
     {
         vec->data = NULL;
         return vec;
@@ -156,7 +154,7 @@ tmpl_Create_Empty_UShortVector(unsigned long int length)
     if (vec->data == NULL)
     {
         vec->error_occurred = tmpl_True;
-        vec->length = 0UL;
+        vec->length = (size_t)0;
         vec->error_message = tmpl_strdup(
             "Error Encountered: libtmpl\n"
             "\tFunction Name: tmpl_Create_Empty_UShortVector\n\n"
